feat: Add target for PowerSlider and update UI images

+++ Assets/Scripts/GameEngine/IEventListener.cs
@@ -1,4 +1,4 @@
 public interface IEventListener
 {
-    void EventUpdate(EventEnum eventEnum, int data);
+    void EventUpdate(EventEnum eventEnum, float data);
 }
